requirements.txt
fastapi
pydantic[email]
uvicorn[standard]==0.40.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
python-multipart==0.0.2